+++ src/main/java/com/mycompany/system/controller/StoreStockController.java
@@ -6,70 +6,84 @@ package com.mycompany.system.controller;
 
 import com.mycompany.system.model.business.StoreStock;
 import com.mycompany.system.service.StoreStockService;
-import java.util.List;
-import java.util.Optional;
+import com.mycompany.system.util.JwtTokenUtil;
+import com.nimbusds.jose.JOSEException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+
+import java.text.ParseException;
+import java.util.List;
+import java.util.Optional;
 
 /**
- *
  * @author ro
  */
 @RestController
-@RequestMapping("/store-stock")
+@RequestMapping("/v1/store-stock")
 public class StoreStockController {
 
     @Autowired
     private StoreStockService service;
 
     @GetMapping(produces = {MediaType.APPLICATION_JSON_VALUE,
-        MediaType.APPLICATION_OCTET_STREAM_VALUE, MediaType.APPLICATION_JSON_UTF8_VALUE})
-    public ResponseEntity<List<StoreStock>> getAll() {
+            MediaType.APPLICATION_OCTET_STREAM_VALUE, MediaType.APPLICATION_JSON_UTF8_VALUE})
+    public ResponseEntity<List<StoreStock>> getAll(@RequestHeader(HttpHeaders.AUTHORIZATION) String token) throws ParseException, JOSEException {
+
+        ResponseEntity HTTP_EXCEPTION = JwtTokenUtil.validateToken(token);
+        if (HTTP_EXCEPTION != null) return HTTP_EXCEPTION;
+
         return new ResponseEntity<>(service.getAll(), HttpStatus.OK);
     }
-    
-     
+
+
     @PostMapping("/stock")
-    public ResponseEntity<StoreStock> findByProductIdAndStoreId(
-            @RequestParam(name= "productId") int productId, @RequestParam(name= "storeId") int storeId){
-        
+    public ResponseEntity<StoreStock> findByProductIdAndStoreId(@RequestHeader(HttpHeaders.AUTHORIZATION) String token,
+                                                                @RequestParam(name = "productId") int productId,
+                                                                @RequestParam(name = "storeId") int storeId) throws ParseException, JOSEException {
+
+        ResponseEntity HTTP_EXCEPTION = JwtTokenUtil.validateToken(token);
+        if (HTTP_EXCEPTION != null) return HTTP_EXCEPTION;
+
         Optional<StoreStock> optional = service.findByProductIdAndStoreId(productId, storeId);
-        
-        if(optional.isPresent()){
+
+        if (optional.isPresent()) {
             return new ResponseEntity<>(optional.get(), HttpStatus.OK);
         }
-        
+
         return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        
-        
     }
 
 
     @PostMapping
-    public ResponseEntity<HttpStatus> create(@RequestBody StoreStock storeStock) {
+    public ResponseEntity<HttpStatus> create(@RequestHeader(HttpHeaders.AUTHORIZATION) String token, @RequestBody StoreStock storeStock) throws ParseException, JOSEException {
+
+        ResponseEntity HTTP_EXCEPTION = JwtTokenUtil.validateToken(token);
+        if (HTTP_EXCEPTION != null) return HTTP_EXCEPTION;
+
         service.save(storeStock);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
     @PatchMapping
-    public ResponseEntity<HttpStatus> update(@RequestBody StoreStock storeStock) {
+    public ResponseEntity<HttpStatus> update(@RequestHeader(HttpHeaders.AUTHORIZATION) String token, @RequestBody StoreStock storeStock) throws ParseException, JOSEException {
+
+        ResponseEntity HTTP_EXCEPTION = JwtTokenUtil.validateToken(token);
+        if (HTTP_EXCEPTION != null) return HTTP_EXCEPTION;
+
         service.update(storeStock);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<HttpStatus> delete(@PathVariable int id) {
+    public ResponseEntity<HttpStatus> delete(@RequestHeader(HttpHeaders.AUTHORIZATION) String token, @PathVariable int id) throws ParseException, JOSEException {
+
+        ResponseEntity HTTP_EXCEPTION = JwtTokenUtil.validateToken(token);
+        if (HTTP_EXCEPTION != null) return HTTP_EXCEPTION;
+
         service.delete(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
